dir added in export functions #27

+++ R/export_rda.R
@@ -5,7 +5,8 @@
 #' @param FUN the function using which the file is to write
 #' @param bucket the name of the bucket, if not set globally
 #' @param data_source the name of the data source, if not set globally. s3, gcs or local
-
+#' @param dir the directory to store intermediate files
+#' @param delete_file logical. to delete the file to be uploaded
 #'
 #' @return No output
 #' @export "export_rda"
@@ -17,7 +18,7 @@
 #' }
 
 export_rda <- function(..., file, FUN = save, data_source = flyio_get_datasource(),
-                        bucket = flyio_get_bucket(data_source)){
+                        bucket = flyio_get_bucket(data_source), dir = flyio_get_dir(), delete_file = TRUE){
   # checking if the file is valid
   assert_that(tools::file_ext(file) %in% c("rda", "Rda","RData"), msg = "Please input a valid path")
   if(data_source == "local"){
@@ -25,8 +26,12 @@ export_rda <- function(..., file, FUN = save, data_source = flyio_get_datasource
     return(invisible(t))
   }
   # a tempfile with the required extension
-  temp <- tempfile(fileext = paste0(".",tools::file_ext(file)))
-  on.exit(unlink(temp))
+  if(isTRUE(delete_file)){
+    temp <- tempfile(fileext = paste0(".",tools::file_ext(file)), tmpdir = dir)
+    on.exit(unlink(temp))
+  } else {
+    temp <- paste0(dir, "/", basename(file))
+  }
   # loading the file to the memory using user defined function
   file = gsub("\\/+","/",file)
   FUN(..., file = temp)
