+++ R/export_shp.R
@@ -6,6 +6,8 @@
 #' @param dsnlayerbind if the FUN needs dsn and layer binded or not
 #' @param data_source the name of the data source, if not set globally. s3, gcs or local
 #' @param bucket the name of the bucket, if not set globally
+#' @param dir the directory to store intermediate files
+#' @param delete_file logical. to delete the file to be uploaded
 #' @param ... other parameters for the FUN function defined above
 #' @export "export_shp"
 #' @return output of the FUN function if any
@@ -19,7 +21,7 @@
 
 
 export_shp <- function(obj, pathshp, FUN = rgdal::writeOGR, dsnlayerbind = F, data_source = flyio_get_datasource(),
-                        bucket = flyio_get_bucket(data_source), ...){
+                        bucket = flyio_get_bucket(data_source), dir = flyio_get_dir(), delete_file = TRUE, ...){
   filename = basename(pathshp)
   layer = gsub(paste0("\\.",tools::file_ext(pathshp),"$"), "", filename)
   dsn = gsub(paste0(filename,"$"),"", pathshp)
@@ -51,8 +53,8 @@ export_shp <- function(obj, pathshp, FUN = rgdal::writeOGR, dsnlayerbind = F, da
   # downloading the file
   for(i in shpfiles){
     # a tempfile with the required extension
-    temp <- paste0(tempdir(), "/", i)
-    on.exit(unlink(temp))
+    temp <- paste0(dir, "/", i)
+    if(isTRUE(delete_file)){on.exit(unlink(temp))}
     # uploading the file
     dsnlayer_i = gsub(paste0("\\.",tools::file_ext(dsnlayer),"$"), "", dsnlayer)
     downlogical = export_file(localfile = temp, bucketpath = paste0(dsnlayer_i, ".", tools::file_ext(i)),
